+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - feat: Expose a function to retrieve the URL of the CSP endpoint (#1378)
 - feat: Add support for Dynamic Sampling (#1360)
   - Add `segment` to `UserDataBag`
-  - Add `TransactionSource`, to set information about the transaction name
+  - Add `TransactionSource`, to set information about the transaction name via `TransactionContext::setSource()` (#1382)
   - Deprecate `TransactionContext::fromSentryTrace()` in favor of `TransactionContext::fromHeaders()`
 
 ## 3.8.1 (2022-09-21)